+++ packages/validation/src/index.ts
@@ -1,4 +1,5 @@
 export * from './account'
+export * from './activity'
 export * from './document'
 export * from './pagination'
 export * from './supplier'